refactor
Extract the 'app' folder from the 'src' to the root folder.
This is a first step towards implementing 'clean-architecture' design pattern for this project. The 'src' folder is going to be used for layering the software's logic. The 'app' folder will be in charge of the presentation layer from now on.